+++ examples/partisipa_close/lifecycle/staged_replay.py
@@ -20,7 +20,7 @@ from typing import Any
 
 from django.apps import apps
 
-from django_rakaia.effect_executor import DjangoExecutor
+from django_rakaia import DjangoExecutor
 from rakaia import upcast