+++ examples/partisipa_merge/subproject/merge_replay.py
@@ -24,7 +24,7 @@ from typing import Any
 
 from django.apps import apps
 
-from django_rakaia.effect_executor import DjangoExecutor
+from django_rakaia import DjangoExecutor
 from rakaia import upcast